+++ runtime/stringbuf.c
@@ -357,7 +357,7 @@ rsRetVal rsCStrAppendInt(cstr_t *pThis, long i)
 
 	rsCHECKVALIDOBJECT(pThis, OIDrsCStr);
 
-	CHKiRet(srUtilItoA((char*) szBuf, sizeof(szBuf), i));
+	CHKiRet(srUtilItoA((char*) szBuf, sizeof(szBuf), (int64_t)i));
 
 	iRet = rsCStrAppendStr(pThis, szBuf);
 finalize_it:
